nest all non-anchor nav (menu) li text in spans
Must do for Foundation so that the text in nav (menu, per most recent Foundation terminology) items that is not in anchor - handy courtesy of Rails' link_to_unless() - can be selected and formatted the same way the anchor text is.

Generally, it's a good practice, I should have done this right away (avoid text just floating around :).