refactor rendering of /portfolios/.. pages
Make code consistent, easier to read, understand, flow with, maximize utilizing of Rails' conventions.

Define the necessary methods on Portfolio model (subset of the methods defined on Admin::Portfolio model) and invoke the Portfolio model in PortfoliosController (insted of the Admin::Portfolio). Bring the portfolio partial from /admin/portfolios back to /portfolios where it belongs - it should live in the namespace in which it is rendered: /portfolios, not /admin/portfolios/ where it's never rendered.