rename contact path helpers for form_for @inquiry to route properly
The path helpers cannot be named `new_contact` and `contacts` (the _path suffix being supplied by Rails).

Rails infers path names from the class of @inquiry passed to form_for on /contacts/new page to be new_inquiry_path and inquiries_path, and these are the paths Rails looks for in the router.

Alternatively, have a Contact model instead of an Inquiry for the online inquiry form.